chore: resolved commment feedback
